Add support for building on Darwin/OS X.
Update code to separate Linux and Darwin specific flags, and
circumvent macros missing from Apple's development environment.

+++ src/plugins/strategies/libvertex.c
@@ -287,8 +287,7 @@ int vertex_regrid(vertex_t* v)
             break;
 
         case HVAL_REAL:
-            v->term[i] = hrange_real_nearest(&range[i].bounds.r,
-                                             (long)v->term[i]);
+            v->term[i] = hrange_real_nearest(&range[i].bounds.r, v->term[i]);
             break;
 
         case HVAL_STR:
